askys-token/requirements.txt
Flask==3.0.3
google-auth==2.37.0
gunicorn==22.0.0
requests==2.32.3
Werkzeug==3.0.3
